Change async backoff config storage

+++ src/channel.rs
@@ -444,7 +444,7 @@ impl<T> ChannelShared<T> {
     }
 
     #[inline]
-    pub(crate) fn detect_async_backoff_tx(&self) -> i8 {
+    pub(crate) fn detect_async_backoff_tx(&self) -> u16 {
         // Async parameter is determine by runtime,
         // like tokio you might have multiple runtime. So the result should stored in
         // sender and receivers, not in the ChannelShared
@@ -463,7 +463,7 @@ impl<T> ChannelShared<T> {
     }
 
     #[inline]
-    pub(crate) fn detect_async_backoff_rx(&self) -> i8 {
+    pub(crate) fn detect_async_backoff_rx(&self) -> u16 {
         // Async parameter is determine by runtime,
         // like tokio you might have multiple runtime. So the result should stored in
         // sender and receivers, not in the ChannelShared
